GameCharacterService -> 스킬 사용 성공 Test Code 작성

+++ ooptest/src/main/java/com/biginsight/ooptest/domain/Skill.java
@@ -21,7 +21,7 @@ public class Skill {
     @Enumerated(EnumType.STRING)
     private CharacterSpecies characterSpecies;  // 해당 스킬 사용이 가능한 캐릭터 종족
     private String name;            // 스킬명
-    private Integer requiredMp;     // 소요 mp
+    private Float requiredMp;     // 소요 mp
     private Integer requiredLevel;  // 필요 레벨
     private String effect;          // 스킬 효과
 }
